tolto il simulatore e creata tabella snr pd soglia nel main

+++ src/radio/detector/EnergyDetector.java
@@ -59,7 +59,7 @@ public class EnergyDetector {
 				count++;
 			}
 		}
-		return count;
+		return (double) count;
 	}
 
 	/**
